+++ src/ht_etl/domain/uniformized_data/gene_expressions.py
@@ -7,8 +7,8 @@ Build uniformized data object for every dataset.
 # third party
 
 # local
-from src.ht_etl.domain.uniformized_data.uniformized_base import Base
-from src.ht_etl.domain.uniformized_data.domain.gene_expression import GeneExpression
+from ht_etl.domain.uniformized_data.uniformized_base import Base
+from ht_etl.domain.uniformized_data.domain.gene_expression import GeneExpression
 
 
 class GeneExpressions(Base):
