+++ src/ht_etl/domain/uniformized_data/sites.py
@@ -7,9 +7,9 @@ Build uniformized data object for every dataset.
 # third party
 
 # local
-from src.ht_etl.domain.uniformized_data.uniformized_base import Base
-from src.ht_etl.domain.uniformized_data.domain.site import Site
-from src.libs import utils
+from ht_etl.domain.uniformized_data.uniformized_base import Base
+from ht_etl.domain.uniformized_data.domain.site import Site
+from libs import utils
 
 
 class Sites(Base):
