+++ src/ht_etl/domain/uniformized_data/tts.py
@@ -7,8 +7,8 @@ Build uniformized data object for every dataset.
 # third party
 
 # local
-from src.ht_etl.domain.uniformized_data.uniformized_base import Base
-from src.ht_etl.domain.uniformized_data.domain.tts import TTS
+from ht_etl.domain.uniformized_data.uniformized_base import Base
+from ht_etl.domain.uniformized_data.domain.tts import TTS
 
 
 class TTSs(Base):
